+++ Polytope_distance_d/include/CGAL/Polytope_distance_d.h
@@ -386,10 +386,10 @@ public:
   // access to point sets
   int  ambient_dimension( ) const { return d; }
     
-  int  number_of_points( ) const { return p_points.size()+q_points.size();}
+  int  number_of_points( ) const { return static_cast<int>(p_points.size()+q_points.size());}
     
-  int  number_of_points_p( ) const { return p_points.size(); }
-  int  number_of_points_q( ) const { return q_points.size(); }
+  int  number_of_points_p( ) const { return static_cast<int>(p_points.size()); }
+  int  number_of_points_q( ) const { return static_cast<int>(q_points.size()); }
     
   Point_iterator  points_p_begin( ) const { return p_points.begin(); }
   Point_iterator  points_p_end  ( ) const { return p_points.end  (); }
@@ -401,10 +401,10 @@ public:
   int
   number_of_support_points( ) const
   { return is_finite() ? 
-      p_support_indices.size()+q_support_indices.size() : 0; }
+      static_cast<int>(p_support_indices.size()+q_support_indices.size()) : 0; }
     
-  int  number_of_support_points_p() const { return p_support_indices.size();}
-  int  number_of_support_points_q() const { return q_support_indices.size();}
+  int  number_of_support_points_p() const { return static_cast<int>(p_support_indices.size());}
+  int  number_of_support_points_q() const { return static_cast<int>(q_support_indices.size());}
     
   Support_point_iterator
   support_points_p_begin() const
@@ -574,8 +574,8 @@ public:
   insert( InputIterator1 p_first, InputIterator1 p_last,
 	  InputIterator2 q_first, InputIterator2 q_last)
   { 
-    int old_r = p_points.size();
-    int old_s = q_points.size();
+    int old_r = static_cast<int>(p_points.size());
+    int old_s = static_cast<int>(q_points.size());
     p_points.insert( p_points.end(), p_first, p_last);
     q_points.insert( q_points.end(), q_first, q_last);
     set_dimension();
@@ -590,7 +590,7 @@ public:
   void
   insert_p( InputIterator p_first, InputIterator p_last)
   { 
-    int old_r = p_points.size();
+    int old_r = static_cast<int>(p_points.size());
     p_points.insert( p_points.end(), p_first, p_last);
     set_dimension();
     CGAL_optimisation_precondition_msg
@@ -603,7 +603,7 @@ public:
   void
   insert_q( InputIterator q_first, InputIterator q_last)
   { 
-    int old_s = q_points.size();
+    int old_s = static_cast<int>(q_points.size());
     q_points.insert( q_points.end(), q_first, q_last);
     set_dimension();
     CGAL_optimisation_precondition_msg
@@ -679,13 +679,13 @@ private:
     if ( ( p_points.size() == 0) || ( q_points.size() == 0)) return;
         
     // construct program
-    unsigned int n = 2 * d + p_points.size() + q_points.size();
-    unsigned int m = d + 2;
+    int n = 2 * d + static_cast<int>(p_points.size() + q_points.size());
+    int m = d + 2;
     CGAL_optimisation_precondition (p_points.size() > 0);
     QP qp (n, m, 
 	   A_iterator 
 	   (boost::counting_iterator<int>(0), 
-	    A_matrix (d, da_coord, p_points.begin(), p_points.size(), 
+	    A_matrix (d, da_coord, p_points.begin(), static_cast<int>(p_points.size()), 
 			 q_points.begin())),
 	   B_iterator (boost::counting_iterator<int>(0), B_vector (d)), 
 	   R_iterator (CGAL::EQUAL), 
@@ -699,7 +699,7 @@ private:
     CGAL_optimisation_assertion(solver->status() == QP_OPTIMAL);
     // compute support and realizing points
     ET  et_0 = 0;
-    int r = p_points.size();
+    int r = static_cast<int>(p_points.size());
     p_coords.resize( ambient_dimension()+1);
     q_coords.resize( ambient_dimension()+1);
     std::fill( p_coords.begin(), p_coords.end(), et_0);
